+++ src/util/io.rs
@@ -6,13 +6,13 @@ pub struct LimitErrorReader<R> {
 }
 
 impl<R: Reader> LimitErrorReader<R> {
-    pub fn new(r: R, limit: uint) -> LimitErrorReader<R> {
+    pub fn new(r: R, limit: usize) -> LimitErrorReader<R> {
         LimitErrorReader { inner: LimitReader::new(r, limit) }
     }
 }
 
 impl<R: Reader> Reader for LimitErrorReader<R> {
-    fn read(&mut self, buf: &mut [u8]) -> io::IoResult<uint> {
+    fn read(&mut self, buf: &mut [u8]) -> io::IoResult<usize> {
         self.inner.read(buf).map_err(|e| {
             if e.kind == io::EndOfFile && self.inner.limit() == 0 {
                 io::IoError {
